New feature: Added install script and Moonraker setup:
Install script will install additional software packages for this
extension and install the extension into Klipper's extras folder.

Added instructions on how to use the install script to the README file.

Supplied the moonracker  update manager section so updates of this extension can be automatically performed.

Fixed a typo in the install script.