Fix enums.py Class PythonVersion
Corrected Class PythonVersion to reflect the correct numbers

In scripts\install.py file corrected line 217 so that the MainMenu will reflect the correct python version.

Add a call to line 289 SaveRequirementsFile so that moonraker's update manager will have to correct requirements.txt file when performing a update for this extension.

Added a install.sh file so that moonraker update manager can refresh the symbolic links when an update occurs.

Change the scripts\utils\custom\extended_macro.py to allow symbolic links to the klipper/extras folder because
most users will use a symbolic link instead of a hard link to the extras folder because that way all the extension files
stay in their own Github repo directories and do not add files
to extras folder because when Klipper needs to be re-cloned the extension files will be lost in the extras folder if a hard link is used.

Removed unnecessary spaces in all the files.

fix moonraker updater in README file.

corrected the path to install.sh for moonraker.

fix delayed_gcode to work in python3 envir