Python version was alway None
In the config.py file, the python version would always return None value

In a python3 virtual environment the call to os.path.realpath(py3_path)
returns a value of "/usr/bin/python3.9" which will never be equal to
self.PythonEnvBinary which is equal to "/home/pi/klippy-env/bin/python"

The purpose of property PythonVersion is to return the number 2 or
number 3 based on which python is installed.

The install.py would stall out on the Input() call on line 82.  This
call to Input() is not needed because the
user has already answered the question about
"Do you want moonraker to update automatically?"